docs(typography): add more usage stories
